requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.10.8
click==7.1.2
gast==0.5.3
grpcio==1.43.0
h5py==3.6.0
hyperlpr==0.0.2
importlib-metadata==4.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
mock==4.0.3
numpy==1.21.5
opencv-python==3.4.9.31
Pillow==9.0.1
protobuf==3.19.4
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.2
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
scipy==1.7.3
six==1.16.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typing_extensions==4.0.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.7.0
